requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-jose
passlib[bcrypt]
python-multipart